chat-app
This project was a requirement for my undergraduate Computer Networking sessional. The main purpose of this project is to get an idea of java socket programming and who applications of different pc can communicate over same local area network. Now our target is to build a good documented and well organized chat app in java so that beginners can get started easily. Kindly come forward to make this project flourish.